food-delivery-mcp/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
uiautomator2>=3.0.0
pyyaml>=6.0
httpx>=0.25.0
pydantic>=2.5.0

